+++ bk2ch05p194touches/ch18p533touches/MyView3.m
@@ -13,7 +13,7 @@
     // be undecided
     self->_decidedTapOrDrag = NO;
     // prepare for a tap
-    int ct = [(UITouch*)touches.anyObject tapCount];
+    NSUInteger ct = [(UITouch*)touches.anyObject tapCount];
     if (ct == 2) {
         [NSObject cancelPreviousPerformRequestsWithTarget:self
                                                  selector:@selector(singleTap)
@@ -59,7 +59,7 @@
 - (void) touchesEnded:(NSSet *)touches withEvent:(UIEvent *)event {
     if (!self->_decidedTapOrDrag || !self->_drag) {
         // end for a tap
-        int ct = [(UITouch*)touches.anyObject tapCount];
+        NSUInteger ct = [(UITouch*)touches.anyObject tapCount];
         if (ct == 1)
             [self performSelector:@selector(singleTap) withObject:nil 
                        afterDelay:0.3];
